Move Github Issues Test Units
