+++ test/issues/github/github_issue_16.e2e.ts
@@ -2,8 +2,8 @@ import supertest = require("supertest");
 
 import {
     connect, drop, newUser, addCategoryAndRegexp
-} from "../helpers/database";
-import { init } from "../helpers/server";
+} from "../../helpers/database";
+import { init } from "../../helpers/server";
 import auth = require("@db/auth");
 
 /**
