+++ test/issues/github/github_issue_28.e2e.ts
@@ -2,8 +2,8 @@ import supertest = require("supertest");
 
 import {
     connect, drop, newUser
-} from "../helpers/database";
-import { init } from "../helpers/server";
+} from "../../helpers/database";
+import { init } from "../../helpers/server";
 import auth = require("@db/auth");
 
 /**
